+++ src/so_vits_svc_fork/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "3.6.1"
+__version__ = "3.10.3"
 
 from .logger import init_logger